put changes in core facet-manager
https://coveord.atlassian.net/browse/KIT-2524
